services/processing/requirements.txt
fastapi
uvicorn
celery
redis
openai-whisper
sentence-transformers
faiss-cpu
rank-bm25
nltk
psycopg2-binary